Only download Xcode in the iOS test runner if LUCI_CONTEXT is present
Bug: 362906088
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6030451
Commit-Queue: Will Yeager <wyeager@google.com>
Cr-Commit-Position: refs/heads/main@{#1390349}

+++ ios/build/bots/scripts/test_runner.py
@@ -463,7 +463,7 @@ class TestRunner(object):
 
   def wipe_derived_data(self):
     """Removes the contents of Xcode's DerivedData directory."""
-    if os.path.exists(DERIVED_DATA):
+    if os.path.exists(DERIVED_DATA) and not xcode_util.is_local_run():
       shutil.rmtree(DERIVED_DATA)
       os.mkdir(DERIVED_DATA)
